+++ compiler/extensions/java/runtime/src/zserio/runtime/validation/ValidationBitStreamReader.java
@@ -137,9 +137,9 @@ public final class ValidationBitStreamReader extends ByteArrayBitStreamReader
     {
         // set NaN used by writer (0x7E00) to nanMaskBuffer
         final byte nanHighByte = (byte)0x7E;
-        nanMaskBuffer[startBytePosition] |= (nanHighByte >> startBitOffset);
+        nanMaskBuffer[startBytePosition] |= (byte)(nanHighByte >> startBitOffset);
         if (startBitOffset != 0)
-            nanMaskBuffer[startBytePosition + 1] |= (nanHighByte << (8 - startBitOffset));
+            nanMaskBuffer[startBytePosition + 1] |= (byte)(nanHighByte << (8 - startBitOffset));
     }
 
     private interface MaskBufferAction
@@ -162,7 +162,7 @@ public final class ValidationBitStreamReader extends ByteArrayBitStreamReader
             // --------|------|------|------|------|------|------|------|------|------|
             // mask    | 0x00 | 0x01 | 0x03 | 0x07 | 0x0F | 0x1F | 0x3F | 0x7F | 0xFF |
             final int mask = (1 << numBits) - 1;
-            maskBuffer[bytePosition] |= mask << (8 - bitOffset - numBits);
+            maskBuffer[bytePosition] |= (byte)(mask << (8 - bitOffset - numBits));
         }
 
         @Override
@@ -188,7 +188,7 @@ public final class ValidationBitStreamReader extends ByteArrayBitStreamReader
             // --------|------|------|------|------|------|------|------|------|------|
             // mask    | 0xFF | 0xFE | 0xFC | 0xF8 | 0xF0 | 0xE0 | 0xC0 | 0x80 | 0x00 |
             final int mask = ~((1 << numBits) - 1);
-            maskBuffer[bytePosition] &= mask << (8 - bitOffset - numBits);
+            maskBuffer[bytePosition] &= (byte)(mask << (8 - bitOffset - numBits));
         }
 
         @Override
